+++ ndsampler/abstract_frames.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 Fast access to subregions of images.
 
@@ -12,10 +11,8 @@ TODO:
         - [X] Use as COG if input file is a COG
         - [X] Convert to COG if needed
 """
-from __future__ import absolute_import, division, print_function, unicode_literals
 import numpy as np
 import ubelt as ub
-import six
 import copy
 import warnings
 from os.path import exists, join
@@ -183,7 +180,7 @@ class Frames(object):
                 backend = 'cog'
             else:
                 backend = 'npy'
-        if isinstance(backend, six.string_types):
+        if isinstance(backend, str):
             backend = {'type': backend, 'config': {}}
 
         backend_type = backend.get('type', None)