+++ ndsampler/utils/util_lru.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 import collections
 import ubelt as ub